Add gctweaks module.
After doing a lot of configuration, I ended up creating a tweaks module for the site, to fix a few bits of weirdness. I also included some code to override the user picture setting. It's currently hard-coded to a field, but I might make some settings for it. I also altered how breadcrumbs work for user groups, to make it a little easier to work with.